resume/requirements.txt
Flask
WeasyPrint
flask==2.0.1
pdfkit==1.0.0
python-pptx==0.6.21
Werkzeug==2.0.1
